+++ src/helpers/Utils.ts
@@ -16,6 +16,7 @@ import { SolutionComponent } from '../entities/SolutionComponent';
 import { ComponentType } from '../entities/ComponentType';
 import { promises as fsPromises } from 'fs'; 
 import { CanvasApp } from '../entities/CanvasApp';
+import { getOutputChannel, outputHttpLog, outputHttpResult } from '../extension';
 export class Utils {
 	
     static async postWithReturnArray<T>(url: string, convert: (ti: any) => T, sort: ((t1: T, t2: T) => number) | undefined, filter: ((t1: T) => boolean) | undefined, content: any | undefined, contentType: string | undefined, bearerToken? : string | undefined): Promise<T[]> {
@@ -26,9 +27,13 @@ export class Utils {
         if (bearerToken !== undefined) {
             headers.Authorization = `Bearer ${bearerToken}`;
         }
+
+        outputHttpLog(`    GET ${url}`);
         var response = await axios.default.post(url, content, {
             headers: headers
         });
+        outputHttpResult(response);
+
         var tis = [];
         if (response.data.value === undefined) {
             tis = response.data
@@ -54,9 +59,13 @@ export class Utils {
         if (bearerToken !== undefined) {
             headers.Authorization = `Bearer ${bearerToken}`;
         }
+        
+        outputHttpLog(`    GET ${url}`);
         var response = await axios.default.get(url,{
             headers: headers
         });
+        outputHttpResult(response);
+        
         var tis = [];
         if (response.data.value === undefined) {
             tis = response.data
@@ -163,12 +172,17 @@ export class Utils {
 
     static async executeChildProcess(cmd: string, onSuccess?: Action<any> | undefined, onError?: Action<any> | undefined): Promise<boolean> {
         try {
+            getOutputChannel().show(true);
+            getOutputChannel().append(`\n\nRUN: ${cmd}\n`);
+                        
             const result = await new Promise((resolve, reject) => {
                 const cp     = require('child_process');
                 cp.exec(cmd, (error: any, stdout: string, stderr: string) => {
                     if (error) {
+                        getOutputChannel().append(`${error}\n`);
                         reject(error);
                     } else {
+                        getOutputChannel().append(`${stdout}\n`);
                         resolve(stdout); 
                     }
                 });
@@ -209,8 +223,7 @@ export class Utils {
         await copyRecursiveSync(sourceFolder, targetFolder);
     } 
     
-    static async getSourceFileUtilityCommandLine(args?: string): Promise<string> {
-        const binPath = await Utils.getSourceFileUtility();
+    static async getToolsCommandLine(binPath: string, args?: string): Promise<string> {
         const os = require('os');
         const fs = require('fs');
         if (fs.existsSync(binPath) && `${os.platform}`.toLowerCase() !== "win32") {
@@ -228,7 +241,15 @@ export class Utils {
         return `${binPath} ${args}`;
     }
 
-    static async getSourceFileUtility(): Promise<string> {
+    static async getPASopaUtilityCommandLine(args?: string): Promise<string> {
+        return await this.getToolsCommandLine(await Utils.getPASopaPath(), args ?? "");    
+    }
+
+    static async getSolutionPackerCommandLine(args?: string): Promise<string> {
+        return await this.getToolsCommandLine(await Utils.getSolutionPackerPath(), args ?? "");    
+    }
+
+    static async getPASopaPath(): Promise<string> {
         const os = require('os');
         const fs = require('fs');
         let binPath = Settings.sourceFileUtility();
@@ -253,17 +274,59 @@ export class Utils {
         return Settings.sourceFileUtility();
     }
 
+    static async getSolutionPackerPath(): Promise<string> {
+        const os = require('os');
+        const fs = require('fs');
+        let binPath = Settings.coreToolsSolutionPackager();
+        if (fs.existsSync(binPath)) { return binPath; }
+        switch (`${os.platform}`.toLowerCase()) {
+            // Windows
+            case "win32":  binPath = path.join(path.dirname(__filename), "..", "..", "bin/windows/CoreTools/SolutionPackager.exe"); break;
+            
+            // Mac-OS
+            case "macos":
+            case "darwin": binPath = path.join(path.dirname(__filename), "..", "..", "bin/macos/CoreTools/SolutionPacker.dll");   break;
+            
+            // Linux
+            case "linux":
+            case "freebsd":
+            case "openbsd": 
+            case "ubuntu":           
+            default:       binPath = path.join(path.dirname(__filename), "..", "..", "bin/ubuntu/CoreTools/SolutionPacker");  break;            
+        }
+        
+        if (fs.existsSync(binPath)) { return binPath; }
+        return Settings.coreToolsSolutionPackager();
+    }
+
+
     /**
-     * Check the Source file Utility for Pack & Unpack.
+     * Check the source file utility (PASopa) for Pack & Unpack PowerApps.
      * @returns success
      */
-    static async checkSourceFileUtility(): Promise<boolean> {
-		const sourceFileUtility = await Utils.getSourceFileUtilityCommandLine();
+    static async checkPASopaTool(): Promise<boolean> {
+		const sourceFileUtility = await Utils.getPASopaUtilityCommandLine();
         var success = await Utils.executeChildProcess(sourceFileUtility, () => {}, () => {});
         if (success) {
             return true;
         } else {
             vscode.window.showErrorMessage(new vscode.MarkdownString(`The configured Power Apps Source File Pack and Unpack Utility '${sourceFileUtility}' was not found. Please download, compile and setup the tool from https://github.com/microsoft/PowerApps-Language-Tooling`).value);
+            return false;
+        }
+	}
+
+    /**
+     * Check the Core Tools Solution Packer for Pack & Unpack solutions.
+     * @returns success
+     */
+     static async checkSolutionPackerTool(): Promise<boolean> {
+        const fs = require('fs');
+		const solutionPackerUtility = await Utils.getSolutionPackerPath();
+        var success = fs.existsSync(solutionPackerUtility);
+        if (success) {
+            return true;
+        } else {
+            vscode.window.showErrorMessage(new vscode.MarkdownString(`The configured CrmSdk CoreTools Solution-Packer tool '${solutionPackerUtility}' was not found. Please download the tool from https://www.nuget.org/packages/Microsoft.CrmSdk.CoreTools`).value);
             return false;
         }
 	}
@@ -282,7 +345,7 @@ export class Utils {
     }
 
     static onSuccess(result: any) {
-        vscode.window.showInformationMessage(`${result}`);;
+        vscode.window.showInformationMessage(`${result}`);
     }
 
     static onError(err: any) {